got Async working, move file processing to another service
