+++ src/main/java/uk/gov/ons/fwmt/job_service/ApplicationConfig.java
@@ -11,7 +11,9 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.context.annotation.Bean;
+import org.springframework.core.task.TaskExecutor;
 import org.springframework.scheduling.annotation.EnableAsync;
+import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
 import org.springframework.web.client.RestTemplate;
 
 /**
@@ -50,6 +52,17 @@ public class ApplicationConfig {
       //storageService.deleteAll();
       //storageService.init();
     };
+  }
+
+  @Bean(name="processExecutor")
+  public TaskExecutor workExecutor() {
+    ThreadPoolTaskExecutor threadPoolTaskExecutor = new ThreadPoolTaskExecutor();
+    threadPoolTaskExecutor.setThreadNamePrefix("Async-");
+    threadPoolTaskExecutor.setCorePoolSize(3);
+    threadPoolTaskExecutor.setMaxPoolSize(3);
+    threadPoolTaskExecutor.setQueueCapacity(600);
+    threadPoolTaskExecutor.afterPropertiesSet();
+    return threadPoolTaskExecutor;
   }
 
   @Bean
