+++ src/main/java/uk/gov/ons/fwmt/job_service/controller/JobServiceController.java
@@ -34,6 +34,7 @@ import java.io.IOException;
 @Slf4j
 @RestController
 public class JobServiceController {
+
   private final JobService jobService;
 
   @Autowired
