+++ src/main/java/uk/gov/ons/fwmt/job_service/service/JobProcessService.java
@@ -1,0 +1,4 @@
+package uk.gov.ons.fwmt.job_service.service;
+
+public interface JobProcessService {
+}
